pyproject.toml
[tool.poetry]
name = "shinygpt"
version = "0.1.0"
description = ""
authors = ["Chris Busillo <info@shinycomputers.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
pylint = "^3.0.3"
black = "^23.12.1"
openai = "^1.9.0"
python-dotenv = "^1.0.0"
stdlib-list = "^0.10.0"
docker = "^7.0.0"
transformers = "^4.37.1"
torch = "^2.1.2"
channels = "^4.0.0"
uvicorn = { extras = ["standard"], version = "^0.27.0" }
django-cors-headers = "^4.3.1"
channels-redis = "^4.2.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
django-stubs = "^4.2.7"
docker-stubs = { git = "https://github.com/rdozier-work/docker-stubs.git" }
watchdog = { extras = ["watchmedo"], version = "^3.0.0" }
types-requests = "^2.31.0.20240106"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
